finish overlay icon to scroll to top
Also animate the scrolling of the webView, and pressing the icon.
Move the icon to the top, less intrusive.

Here's what it does:
When the user has scrolled the page for more than 2000 pixels
vertically, we show an overlay icon while scrolling up (and a bit
thereafter). Tapping on this icon moves up the page quickly. This is
useful after reading a long page, and one wants to quickly move up to
the header to enter a new address.