don't smooth image while scaling the webpage during zoom
I'm not sure if this is *really* needed. Rendering is disabled, so the
webpage won't do layout updates which makes for the biggest performance
gain already. The image still seems scaled smoothly though, so we
disable that as well and just scale the frames during zoom animations
coarsely.

There's no visible performance gain on my laptop, but let's try anyway
if it squeezes out a few more FPS on devices.