Font size setting from config
We now read an adjustment value to the fontsize from the config, which
will be added to the pixelsize. We also use the 4 point offset from the
last patch, to get a sensible default value.

The browser settings module gets a slider that allows adjusting the
fontsize.

All works nicely. :)