refactor pages to be shown as pages, not layers
By pushing pages into the stack, we can swipe out the dialog
page (Settings, Bookmarks, ...) if we wish.

The browser page (index=0) is made to fill column width with all
other pages required to specify 'Kirigami.ColumnView.fillWidth:
false' in their properties. This ensures that the width of the
pages is adjusted correctly in narrow and wide modes.

The browser page is pinned as well. So, in wide mode, all pushed
pages will be shown on the top of each other.

When the browser page becomes current, all other pages are popped
out from the stack.